backend/requirements/base.txt
Django==5.2
djangorestframework==3.16.0
django-cors-headers==4.7.0
whitenoise==6.9.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
celery==5.5.2
redis==5.2.1
drf-spectacular==0.28.0
django-allauth==65.7.0
djangorestframework-simplejwt==5.5.0
python-dateutil==2.9.0
django-environ==0.12.0
Pillow==11.2.1
